Bugfixes
Fixed a bug with URL input checking. Will now check if entered URL begins with either "http://" or "https://" and if neither is found, it will prepend "http://" to the enetered URL.
Fixed a bug when checking if an entry exists for an entered longUrl. Will now check the existing entries based on the generated shortUrl, rather than the longUrl entered in the form. This is necessary due to the possible prepending of "http://" onto the enetered string.